refactor: split monitoring administration routes
